Converted Cell, Row and Table to be immutable, and added equality methods and associated tests.
Implemented methods for Row and Table to get the number of cells and rows respectively.

Added a get row method to Table.

The Extractor now returns a map of documents and tables, and also has been cleaned and optimized to use TabulaConverter.

A defined extraction flow has been added to Extractor (but not yet fully implemented, pending Stencil, Window and Constraint remodelling)